Fix bigint in dev DB's (#1197)
D's are bigints in dev DB's, whereas they are integers in the prod DB. This causes some queries to fail in dev, and may cause us to write queries in the future that will succeed in dev but fail in prod. Example error from filtered session/error queries: ERROR: operator does not exist: bigint[] && integer[] (SQLSTATE 42883).

I'm not sure how this bigint/integer discrepancy happened, one theory is that it depends on the architecture the machine is running on (we are not explicit on the int type in GORM).

Fix is to explicitly set the type to SERIAL (which is an auto-incrementing integer in postgres).