HIG-360: Release Live Mode MVP for Highlight employees (#1655)
Unprocessed sessions will autoplay in live mode, with a button to toggle in and out. Playback will typically be 15s behind real-time, to reduce jittering. Pausing or leaving a comment will toggle you out of live mode, but you can toggle back in when ready.

To be done in separate PR's:

    Better UI after we've played around with this a bit - just did something crappy for now
    Poll for incremental data, rather than re-fetching the entire payload each time
    Handling when a live session ends
    Real-time alerts