Fix live mode and autoplay interaction (#1886)
When we run out of data to play in live mode (for example, if there is a delay in receiving data, or if the user was simply idle), we have to stop the player at that timestamp, otherwise it will get confused about what to render when the data eventually does come in. This is actually the same logic as what happens when we reach the end of a non-live session - the player will stop when the time runs to the end of the available data.

However, when autoplay is on and we reach state 'SessionEnded', we will play the next session. We don't want this for live mode, so we can use the state 'Paused' instead, which is actually more semantically correct.