Send beacon on session end (#2095)
## Use the 'visibilitychange' event instead of 'beforeunload', as recommended by MDN. This has two advantages:

1. Supported on more browsers, especially mobile, for which 'beforeunload' may not be fired.
2. It is [non-blocking on navigation](https://www.notion.so/runhighlight/Non-Blocking-SendBeacon-Quick-Notes-0defe6519ddc4fcba7703ed2dc981666), for better user experience.

## There are some caveats:

1. 'visibilitychange' is called on tab change, app background, etc, so it can be fired multiple times.
2. 'sendBeacon' is fire-and-forget, so we get no server response and no retry logic if the upload fails.
3. 'sendBeacon' has a size limit of 64kb on most browsers.

## How we will address these caveats:

If we lose the last payload at the very end of the session, that's okay, what we want to avoid is losing payloads in the middle and corrupting the rest of the session. We can do that as follows.

1. When sending a beacon payload, do not delete it on the client. PushPayload will upload the same payload again (with usual retry logic) if the tab is resumed, so we won't lose data in the middle.
2. On the backend, we will mark all events, errors, resources and messages as 'is_beacon' if they were received from a sendbeacon. Whenever we receive a payload, we will delete all previous sendBeacon payloads first. This makes sure that we don't have any duplicate data.

There are some small race conditions remaining. Firstly, because we get no response for 'sendBeacon', we can't guarantee ordering between 'sendBeacon' and 'pushPayload'. Secondly, in the current implementation, when receiving payloads, we only delete previous beacons if the session's 'BeaconTime' field is set - there is a write-after-read race condition that can happen if a beacon and another payload is received simultaneously. Both of these issues should be vanishingly rare but can be fixed if they crop up.

## How we will measure success rate:

This PR saves a 'beaconTime' field on the session which is only set if the last push was a beacon (and overwritten to null if the last push was not a beacon). We can use SQL queries in a day or so to see the percentage of processed sessions where the last push was a beacon.