[HIG-2126] quick search should support quick paste
When pasting quickly into the quick search box, we may not have time to
query opensearch for the best-match query.
In this case, fall back to a default query based on `identifier`, or email
if the input is an email.

The result is that a quick paste-enter results in a more sensible
query rather than using the default pre-loaded opensearch recommendation
(which ends up being the most-used query).

Testing: new string test