[HIG-2125] add comment follower notifications
When a comments, tags a user in a comment, or replies to a comment,
track their email or slack mention as a new model (CommentFollower) that
allows notifying them for future updates to the comment thread.
The effect is similar functionality to Slack where notifications propogate
for all participants in a thread, unless a user opts-out. This doesn't
yet support the opt-out functionality but it would be easy to do so by
removing relevant CommentFollower rows.