[HIG-2140] remove space from comment modal
Add a header to all comment modals to stay consistent.
Adjust the wording of to be consistent:
`Add a Comment`
`Attach an Issue`
Keeping `Create a Linear Issue` in the dropdown since the context there
is different, where we are choosing to create a new issue rather than
attach an existing one.