[HIG-2160] track click events as session fields (#2441)
Sets up a session property for clicks on the page. A `clickTarget` and `clickSelector` property are added.
The target is typically the exact element clicked, but the selector allows querying the full path of the object in the tree. With a string contains, you can match it to the selector as desired.