[HIG-2160] track click events as session fields (#2447)
Sets up a session property for clicks on the page. A clickTarget and clickSelector property are added.
The target is typically the exact element clicked, but the selector allows querying the full path of the object in the tree.
With a string contains, you can match it to the selector as desired.

#2441 had an issue in the selector logic causing some UIs to lag (https://highlightcorp.slack.com/archives/C02RAK20S6Q/p1649868986776399).
The new selector logic can cause the UI to lag due to the selector optimization.
For click selector tracking, don't optimize the selector since we want the full string anyways (that can be searched easily
for the relevant selector).