[HIG-2127] add session and error pagination (backend) (#2566)
Adds pagination to backend sessions / error-groups opensearch queries using opensearch from syntax.
Pagination makes it easier for customers to triage and share their searches because a URL will reflect a consitent
set of cards based on the query and page being viewed. Currently, pages may get shifted if the underlying results
change but a future change to set start and end dates on the query will fix the result set.