[HIG-2332] remove uncompressed worker payload writing code
Remove all uncompressed file writing for session processing.
Since we don't store the uncompressed versions in S3 anymore,
we don't need to write them either.

The only caveat is we used the uncompressed file as a buffer for
session events for processing. Switch to the compressed file as the
buffer with no read batching ( read the whole file for processing
rather than trying to read in batches based on some delimiter ).
This shouldn't be a problem because the size of the sessions shouldn't
overwhelm the worker RAM, but if it becomes a concern we could use a
JSON reader ( read one object from the compressed file at a time )
to reduce memory consumption. This would require a bit more overhaul
of the session processing logic.

Testing: local stack with ENABLE_OBJECT_STORAGE=true