remove events_objects id constaint
shadow ID field so automigrate won't try to reapply the pkey index.
we need to drop the ID field constaint and autoincrement because
we run out of values for the integer field.
with events_objects, we don't care about querying by id and drop
rows after they are processed.

do this for other object tables where we risk running out of ID values.

checked that these modified tables are not queried by ID.