process kafka messages serially to fix live mode glitches (#2837)
## Motivation
In live mode, we have a lot of out-of-order issues currently due to out-of-order messages in the `events_objects_partitioned` table.
This occurs because we have out-of-order processing of `PushPayload` messages due to the kafka worker pool that processes incoming messages.
This manifests as rrweb replayer warnings during live mode as well as incorrect-looking playback in live mode.
![image](https://user-images.githubusercontent.com/1351531/181860043-f57b8bd3-52a9-4074-a5f4-3ed85faf5cf0.png)
In this case, the replay looks like it is stuck in a half-rendered state inconsistent with the true DOM.
![image](https://user-images.githubusercontent.com/1351531/181860369-0527d8b6-a7ed-41d6-ac65-a3fdee321340.png)

## Solution
Create a kafka consumer for each message processing goroutine.
A kafka consumer is allocated a slice of partitions. The consumer will always retrieve messages in order from those partitions.
Since each message will be processed synchronously after retrieval, this will guarantee that messages for a given partition are processed in order.

## Validation
Originally, I validated the theorized solution by running locally with 1 processing goroutine. I did not encounter the live mode bugs seen above (though I was consistently seeing them before).
With this refactored code running N consumers, I no longer see live mode replay glitches or console warnings from rrweb.

## Concerns
1. we will now allocate more (N ECS containers * M goroutines ~ 256) kafka consumers part of the group. however this doesn't seem to be an issue since it is recommended to have [as many consumers as partitions](https://stackoverflow.com/questions/55063898/what-is-the-recommended-number-of-consumers-per-group-in-kafka), and our cluster currently easily handles thousands of partitions (currently prod has ~768 allocated partitions).
2. we may have slower message processing because we may not be able to batch kafka message retrieval as well. we can easily revert the change in this case or spin up more ECS containers to handle this.