only send first-time user alert once
if a session gets multiple `H.identify` calls, only send the new-user alert once by checking that this is the first time we are updating the `userIdentifier` field.
this is still prone to a race if the two calls happen simultaneously but is a quick fix for now for multiple calls spaced apart (which is what happens with the segment integration).