improve client full snapshot logic
perform the full snapshot before making the async network request
to upload the previous events to ensure that we
have a full snapshot at the beginning of a future
payload.
the way this code was written prior, rrweb could add events
during the async network upload of the previous payload, before the
full snapshot is added to the events, leading to
a future payload with a full snapshot not at the beginning
of the payload.