use secure id for client public graph requests
changes client to use the secure session id for all
requests so that the initializeSession request
can be made non-blocking. this will help make sure
that all client requests can be asynchronous and
an issue with initializeSession can be retried on the
processing side and will not cause any data to be lost.

ensures the backend is compatible with both the old and
the new client. during this transition, the backend will
support messages with session or secure id, and will
partition them by secure id falling back to regular id.
since a client will continue to be old for an existing
session, this should ensure ordering between messages.

if a client reloads and continues a previous session with
the secure id, there may be some out of ordered messages
for the same session since some will be keyed by id
and some by secure id, but we've had parallel processing
of messages before without much concern (in this case,
the session will have already been initialized). this should
only be applicable for a small percentage of clients
during the rollout.

to support old clients, the backend will continue
to be synchronous for the `initializeSession` graph
until all clients upgrade.