fix invite links not working correctly
new changes in #2937 broke the invite-link functionality because
the index.tsx admin loader would always see an error when trying to
load the admin role.
instead, have the backend return the admin as part of the role query with
no role, allowing other parts of the frontend to redirect to the expected error state.

Testing: local invite link testing with non-highlight second account