add babel-plugin-react-wrapped-display-name
Take this code:
```js
const Component = React.memo(() => <div />);
```

When using `React.memo` or other similar component wrappers, the display name of the component is lost, and shows up as "Anonymous" in devtools and profiler. This plugin adds the display name of the wrapped component to the display name of the wrapper, for better debugging.

```js
const Component = React.memo(() => <div />);
Component.displayName = "Component (Memo)";
```