fix dependencies for docker build (#3222)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

The docker push pipeline is failing. 


![image](https://user-images.githubusercontent.com/58678/197025866-c0f17217-eef1-4bd3-b287-143f32620c44.png)

Looks to be a missing dependency. 

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

`cd backend && go get
github.com/highlight-run/highlight/backend/storage`

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A