fix inconsistent error frequencies due to differing queries (#3390)
## Summary

Because we had `ErrorGroup` objects returned with a mix of influx and
opensearch frequency data, the `error_frequency` field's data was
inconsistent. The new error page would call `GetErrorGroupsOpenSearch`
which returned `ErrorGroup` objects populated with the `error_frequency`
array from influxdb, but the `GetErrorGroup` query would return the same
`ErrorGroup` with `error_frequency` populated from opensearch.

This PR removes the `influxdb` setting since data is now populated in
influx. This ensures that the results are consistent.

## How did you test this change?

Local deploy with the new error page off and on shows no changes to the
error frequency charts.

## Are there any deployment considerations?

No.