add project id to debug why a initialize sessions events are failing to process (#3451)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

We have a backlog of kafka events failing to process. At least some of
these are because the project tied to this session doesn't exist (see
[logs](https://app.datadoghq.com/logs?query=failed%20to%20process%20InitializeSession%20task&cols=host%2Cservice&index=&messageDisplay=inline&stream_sort=time%2Cdesc&viz=stream&from_ts=1668442312797&to_ts=1671034312797&live=true)).

This PR adds additional data to determine if this is user error or
something wrong with our code. If it's the former, we'd like to punt on
reprocessing these messages as they're being unnecessarily retried.

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->