Add ability to snooze errors (#3516)
## Summary

Adds the ability to "snooze" errors. Snoozing temporarily suppresses
alerts on errors.

https://www.loom.com/share/5eb77cd2d3e443ee911b764666c775ec

This includes updates in the following areas:

* Adds a `snoozed_at` column to `error_groups` and pushes it to
opensearch when an error group's status changes.
* Adds the ability to "Snooze" from the error alert state dropdown. This
includes a handful of preset times (1 hour, tomorrow, etc.) and a
datepicker.
* Adds an indicator on search result cards to show that an error is
snoozed.
* Adds logic to suppress alerts when an error group is snoozed. Also
moves the error suppression logic further up the call stack. This
suppresses emails as well as Slack alerts now.
* Exports the `useMenu` hook from `Menu` so you can access/update menu
state outside the component.
* Adds an `onVisibilityChange` prop to `Menu` so we can subscribe to
changes and reset the menu state.

### Architecture Notes

We also considered adding a `SNOOZED` state for `error_groups.state`.
However, we decided not to go this route because we didn't want to have
to add logic for updating the records back to the previous state and we
would still need to figure out the right UX to figure out how to
indicate/display error groups that are snoozed.

The approach in this PR seemed like the easiest path forward, but
possibly should be reconsidered as we get feedback from customers or
develop opinions on the UX we want to provide.

## How did you test this change?

Local click test. Can't test in PR preview because of the backend
changes.

~~I also configured alerts for my project locally and that they were
sending for an error I triggered manually, but were suppressed while the
error was snoozed.~~

**Update:** I was unable to click test locally, but am merging and will
click test in prod.

## Are there any deployment considerations?

There is a database change here, but it's adding a column that is `null`
by default.