Send console messages to clickhouse (#3606)
## Summary

This PR intercepts the flow to process a console message and parses +
sends the data to clickhouse.

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

See the note in #3531 about how to get `console.*` messages to be
emitted in development.

* Verified the data landed in Clickhouse:

![Screenshot 2023-01-25 at 3 02 53
PM](https://user-images.githubusercontent.com/58678/214701716-1d36fe07-e67f-4712-b164-c6f60ee0089b.png)

* Added some light unit tests around parsing

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

* `CLICKHOUSE_*` environment variables were added to Doppler.
* Currently, this goes to my test Clickhouse db for a proof of concept.
I've only enabled it for the Highlight project. I'd like to monitor this
on production data before rolling this out more broadly.
* Because this is in the kafka flow, I opted to log errors such that we
wouldn't reprocess message. You'll notice in the PR that I am
unmarshalling the console message byte array into a struct and I'm not
totally sure if the format will always look like this.