Fix query param rendering via use-query-params (#4259)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

See the revert of #4257. 

Basically, when we call `set*` with use-query-params, the updated value
wasn't triggering a re-render.

I discovered that the fix for this was two things:
* upgrading
[use-query-params](https://github.com/pbeshai/use-query-params/tree/master/packages/use-query-params)
to the next major version (see [migration
plan](https://github.com/pbeshai/use-query-params/blob/master/packages/use-query-params/CHANGELOG.md#migrating-from-v1))
* correctly nesting `QueryParamProvider` under `BrowserRouter`

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

I undid the revert of #4257 and confirmed that updating the logs query
correctly re-renders (triggers a new request to load logs).

It seems like my bug occurs only when we are using a `set*` but most
times we just pluck off the query value and then pipe it into something
else. For example, the `sign_up` param on the login form is simply piped
into some form state:
https://github.com/highlight/highlight/blob/07d3e88369aa29ee538291216508362197e024b6/frontend/src/pages/Login/Login.tsx#L106-L112

I did find some examples of the `set*` in our pagination: 

https://github.com/highlight/highlight/blame/07d3e88369aa29ee538291216508362197e024b6/frontend/src/components/SearchPagination/SearchPagination.tsx#L100-L113

and in the query builder:

https://github.com/highlight/highlight/blob/07d3e88369aa29ee538291216508362197e024b6/frontend/src/pages/ErrorsV2/ErrorQueryBuilder/components/QueryBuilder/QueryBuilder.tsx#L1371-L1380

Doing some click testing around there (back/forward buttons, deep
linking, etc) looks to be working as expected.



## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A