add an endpoint to fetch values for a given key (#4308)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

When the user types into the log input and has a filter selected, we
should provide possible values that might facilliate their search. This
will work similar to datadog's search:

![Screenshot 2023-02-16 at 4 17 49
PM](https://user-images.githubusercontent.com/58678/219509944-3939b12e-8755-413a-9da2-64023930b6d5.png)

This is the backend endpoint. A follow up PR will handle wiring this up
to the frontend.


## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Unit test added

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A