replace fingerprint with generated client id (#4476)
## Summary

Customers have reported finding collisions in the `fingerprint` session
value. Our usage of fingerprint.js limits what heuristics it uses so
there are likely collisions in the ids.

Switch to using the generated clientID as the fingerprint for
consistency in our session search and frontend UI.

On our backend, we expect the fingerprint as a string now that is
hashed, so the change should be entirely compatible.

## How did you test this change?

Local deploy.

## Are there any deployment considerations?

New client version.