Handle backward pagination and permalinking logs (backend) (#4417)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR extends #4319 to add backward pagination and support to
permalink to log. A frontend PR will follow that wires up these changes.

### Backward pagination

This continues with the [Relay
spec](https://relay.dev/graphql/connections.htm#sec-Backward-pagination-arguments)
adding a `before` argument.

Similar to handling the `after` cursor logic:

```sql
WHERE Timestamp <= x
  AND ( Timestamp < x OR UUID < y )
ORDER BY Timestamp DESC, UUID DESC
```

we need to invert the logic to handle going "backwards" ("backward" in
flying quotes because it's going forward in time to get the newer logs)


```sql
WHERE Timestamp >= x
  AND ( Timestamp > x OR UUID > y )
ORDER BY Timestamp ASC, UUID ASC # Note the flipped order here
```

If we simply returned this directly though, the result would not match
the existing descending order, hence, we wrap this in a sub-select to
return the order correctly.

```sql
SELECT * FROM logs_window (
  SELECT * FROM logs
    WHERE Timestamp >= x
      AND ( Timestamp > x OR UUID > y )
    ORDER BY Timestamp ASC, UUID ASC
 AS logs_window
) ORDER BY Timestamp DESC, UUID DESC
```


### Permalinking

The relay spec has no concept of permalinking so we leverage that.

A naive approach would be:
* given a cursor, return the log that matches this cursor
* frontend can make two more requests to get the logs before/after this
cursor

This is obviously unperformant and is pretty annoying for the frontend
to have to make all these requests.

Hence, I chose to get a "window" of logs (see
[stackoverflow](https://stackoverflow.com/questions/71699810/it-is-possible-in-cursor-based-pagination-to-get-the-prev-and-next-cursor-in-the/71738696#71738696))
around this cursor. Specifically loading `Limit / 2` logs before and
after this cursor (`Limit` is 100 so with the log matching, the result
is 101 logs).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Unit tests were added (specifically around ensuring
[PageInfo](https://relay.dev/graphql/connections.htm#sec-undefined.PageInfo)
is correct as the frontend needs this to be accurate to determine if it
should load additional requests).

I also verified this behavior with a simple frontend
(https://github.com/highlight/highlight/pull/4486/files#diff-96c1192170afa8f77ed262c7c42ab64f29d1f31091d907d211563db845369e54R103)
that will be more complete in a follow up PR.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A