hotfix: commandbar editing and search experience (#4536)
## Summary

<!--
Ideally, there is an attached Linear ticket that will describe the
"why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

1. Switches the default from identifier to email and back based on the
command bar query.
2. At the moment editing the query in the command bar is unusable
because of the cursor jumping to the end. This PR fixes this by using a
hook for a `keydown` event listener rather than overwriting onKeyDown on
the Input. Also sneaks in refactoring as a result of the investigation
into the unstable cursor issue.


## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->
1. cmd + k 
2. Editing the query and navigating back and forth
3. Trying out email-like queries

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->
no