Use github actions to run a vercel deploy (#4746)
## Summary

Resolves #4733 

This will only run the preview deployment or prod deployment if
`docs-content/**` or `highlight.io/**` is changed. Allows us to have
full control of whether a deploy happens on a given branch / file path.

Note that we're using `vercel`, not `vercel build`, so all build work is
done in Vercel's build suite for now. @Vadman97 , I figured this was a
bit simpler, but long term, might be more economical to run the build in
github actions.

After we confirm this works w/ a deploy to `main`, we can turn off the
vercel integration and use this exclusively.

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?

Run in CI. 

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->