use log row constructor everywhere (#4854)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

We have a lot of data normalization when using the log row constructor
(`NewLogRow`) and it's options (e.g. `WithSeverityText`). But some code
calls the struct (`LogRow`) directly (specifically the tests) and that
makes our code more brittle.

Hence, this PR is doing a couple things:

* ensuring that all log rows are created with `NewLogRow`
* adjusted the constructor to require a timestamp and a project id
(these are the bare minimum requirements for a log, everything else is
optional and should use the `With*` functions)
* Got rid of `LogRowPrimaryAttrs` in favor of more `With*` functions
(e.g. `WithTraceID`)
* Since timestamp is required, I got rid of `WithTimestamp` (the logic
of truncating to the second is still present)
* Since project id is required, I got rid of `WithProjectIDString`. It
wasn't doing error handling (see inline comment below).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Existing tests should pass and code should compile

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A