Product philosophy edits (#5078)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

Edits to product philosophy page (added in #4997). The big change was
renaming the file from `product-philosphy.md` to
`product-philosophy.md`. The downside of doing that is that it makes the
diff harder to to discern what has changed through the Github UI. Use
this [diff](https://www.diffchecker.com/ST91U36x/) to see all the
changes.

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Visual test

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A