"Request a Demo" fixes (#5033)
## Summary

- Clicking out of the widget should close it consistently now.
- Calendly includes an empty margin on its widget that may block some
clicks
- Close button is now on the top right, and has an "X" icon instead of a
left arrow
- Pressing Esc should close the widget if it's not focused.
- When the widget is focused it seems to capture the key press and
prevents headless from handling it
- Added button to feature pages

~~There's a big issue I have to figure out when it comes to using the
widgets in other pages. When opening the widget, changing page, then
opening it again, the widget will not appear. Calendly's inline widget
assumes that after the div initially loads, it will *remain* on the
page, which doesn't in our case. I think a different widget type may
work better, but I have to test that.~~

I'm using a library `react-calendly` now, that seems to fix the loading
issue

Closes #5017

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->