avoid duplicate errors from error boundary and show dialog by default (#5167)
## Summary

* Disables error reporting to highlight by the error boundary when
highlight is present because highlight will be catching the same error
via the `window.onerror` listener, resulting in duplicate errors. ie.

![image](https://user-images.githubusercontent.com/1351531/235011726-1cacd889-8e03-4130-8faf-3852e7943898.png)
* Changes the ErrorHandler props to default to `showDialog` and add an
optional `noShowDialog` as we've heard from customers that the props are
un-intuitive (the dialog is the whole point of the error boundary to
begin with).


## How did you test this change?

buttons page

![image](https://user-images.githubusercontent.com/1351531/235014279-a08b6144-5dd2-44f8-a25d-f9c9c1563305.png)
only one error recorded

![image](https://user-images.githubusercontent.com/1351531/235014346-abbb9d24-a339-4645-8fc8-1c87b28a89db.png)


## Are there any deployment considerations?

The `showDialog` is kept to keep this change non-breaking. Since we just
released the major version 3.x today, thinking we can still sneak this
in as 3.0.1 though the behavior of the dialog has changed.