Add generated files check for reflame to workflow (#5366)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

We've seen a few instances where generated files in main are out-of-date
and cause reflame deploys to fail. This PR adds a check to make sure
generated files for reflame are up-to-date.

This does add ~30-40s to the `Build Yarn Turborepo` step, which is not
ideal. We can run this as a separate workflow in parallel with `Build
Yarn Turborepo`, but the step depends on `yarn install` so it will
result in double the build minute usage for that step.

## How did you test this change?

I enabled the action in my fork and verified that it fails on
out-of-date generated files:

<img width="614" alt="Screenshot 2023-05-17 at 1 54 46 PM"
src="https://github.com/highlight/highlight/assets/6934200/19d70a79-9644-4b59-9c59-68a436c39a62">

Adding the files got it passing again.

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->