Loading state in right panel (#5403)
## Summary

Added loading state to appear in right panel when navigating through
setting pages and data is still being retrieved. It's not part of the
App loading state since it's just individual setting pages, so I thought
to replace many of the
```
if (loading) {
return null
}
```

with this new flexible loading div. 

https://www.loom.com/share/85552ecf9fba4bcb9954927d30a509ab

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->