remove feedbackWidget option from client (#5413)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR removes the `feedbackWidget` option from the client. This is an
unused feature from way back in the day that nobody really uses.
Note: this is different from the `<ErrorBoundary />` component but they
both use the same endpoint of `addSessionFeedback`.


## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Confirmed `<ErrorBoundary />` feedback still works correctly.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

Since this is a breaking change, I've bumped the major version.