make all updates to opensearch via private graph synchronous (#5531)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

Similar to #5524, this makes all private graph requests that update
opensearch to be synchronous

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Confirmed that an update to opensearch via private graph (e.g. marking
an error group as viewed) updates as expected.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

Will monitor the same as #5524. Private graph traffic is a lot more
negligible than public graph so I don't expect this to have any impact.