Stop minifying generated css again (#5554)
## Summary

This PR reverts back to outputting non-minified generated css for
Reflame, in hopes that it could minimize frequency of merge conflicts in
the generated file.

Previously we had to start minifying because vanilla-extract's esbuild
plugin outputted comments that depend on absolute path, so would differ
between CI and local machines. In this PR we work around that by
manually stripping out those comments with a regex.

Also reverts #5551, since treating the file as binary actually does the
opposite of what we want (minimize frequency of merge conflicts), since
no automatic conflict resolution could ever occur with binary files.

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?

Styles in latest Reflame deployment looks fine. Should be good to go if
the Reflame generated files check passes in CI. Edit: looks like we're
good to go!

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->