wire up backend to fetch error objects for an error group (#5557)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR wires up the backend to support an error instances listview
(internally, the data model is called `error_object`):

![image](https://github.com/highlight/highlight/assets/58678/7aed2fce-b7f5-4be8-bc3f-f73f662796c3)


Similar to what we do on the logs page, the API follows the [relay
cursor spec](https://relay.dev/graphql/connections.htm). We're able to
do the more performant cursor pagination (over offset pagination)
because we will not support page numbers (like we do in the
sessions/errors feed) i.e. it's just a Next and Previous button.

Unlike the logs page, we have an easier job of using the
autoincrementing id as the cursor designation (clickhouse doesn't have
this concept so we have a different approach, see #4319).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Unit tests added

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A