WebSocket client side recording (#5680)
## Summary
Start recording web socket events and pushing them to the backend via
PushPayload. Note that they are currently being accepted by the backend,
but dropped (i.e. not stored in Redis or Postgres)

- Websocket recording will be on by default if `recordHeadersAndBody` is
enabled
- It can be disabled by setting `disableWebSocketEventRecordings: true`
in the `H.init()` options
- The WebSocket Proxy will be used even if recording is disabled since
this needs to be set up before Highlight is initialized with the
settings in order to catch earlier created WebSockets, but it will not
record any events

## How did you test this change?
Click test:
1. Find a live session, and view in Live Mode
- [ ] Confirm that a web socket was created in the Chrome Network Tab to
the private graph
- [ ] Confirm that an `open`, `send`, and `received` event was sent as a
resource in the `PushPayload` mutation
2. View the current session
- [ ] Confirm the WebSocket events are displayed in Highlight's Network
Tab (may need to scroll to find)
- [ ] Confirm the WebSocket events are interspersed in the Network
requests appropriately, sorted by time


https://www.loom.com/share/70ba511c01bc4850863070c70d0b78d2?sid=fbafcd12-c8fe-46db-9cd6-b74151414522

## Are there any deployment considerations?
The npm package has not been updated, so this will not release a new
build. The frontend should be updated to handle WebSocket events before
the package is released

---------

Co-authored-by: Spencer Amarantides <spenceramarantides@Spencers-MacBook-Pro.local>