attach error group secure id to error objects request (#5814)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR attaches the error group secure id when fetching a list of error
objects. This is useful at the time we render the error instances table
(not available yet) when linking out:


https://github.com/highlight/highlight/blob/b35d9614871ed9a498eba365e71ebad0e98ab05e/frontend/src/pages/ErrorsV2/ErrorInstances/ErrorInstancesTable.tsx#L80-L84

and not having to deal with conditional logic. 

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Unit tests

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A