Display websocket messages in right panel (#5881)
## Summary

https://github.com/highlight/highlight/issues/5780

Adapted the new right panel to display websocket messages and details

https://www.loom.com/share/12017f2a0fa142c082b9487f7272e6f4

The message panel is using the Virtuoso, but it's weirdly rendering the
1st element for 1 second then renders the rest as shown in the Loom. All
the data is present, so loading state is not helping. Not sure why this
is happening

## How did you test this change?

Tested out multiple sessions with websockets. Clicked around for the
messages. Verified with Chrome dev tools

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->