cleanup nextjs docs (#5943)
## Summary

Wanted to clean up the next.js docs to make it a bit more concise, as it
is the main
entrypoint for folks getting set up on next.js. We want to make it as
bare-bones as possible
to avoid confusing folks (as is, there are three steps to instrumenting
in the client, api fns, and serverless,
and it can get overwhelming where to start as an end user). Also, the
warning about limitations
seemed a bit strong and discouraging, so I've refactored it to be in the
relevant sections.

@deltaepsilon appreciate the work you did to build these docs. if there
is anything you disagree with and think
we should preserve, please comment!

## How did you test this change?

[Vercel
preview](https://highlight-landing-git-vadim-cleanup-nextjs-docs-highlight-run.vercel.app/docs/getting-started/fullstack-frameworks/next-js).

## Are there any deployment considerations?

No

---------

Co-authored-by: Jay Khatri <jaykhatrimail@gmail.com>