Extract query parsing logic out of logs (#6014)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR extracts out the query parsing logic from logs. The intention
here is that the syntax can be re-used easily without being tied to a
specific table or database.

Furthermore, it becomes incredibly easier to write additional unit tests
without a clickhouse dependency (see `queryparser_test.go`).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

This was already heavily tested in `logs_test.go`. I added additional
tests for sanity.

Confirmed searching for logs and error instances works as expected

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A