Limit Stacktrace to 5000 characters when querying OpenAI (#6155)
## Summary
**Describe the bug**
A customer tried to ask Harold for help on an error that had a
stacktrace > 27k characters (~9000 tokens). OpenAI returns an error when
we use more than the `4097` token max. Frontend errors that use
`MappedStackTrace` are more susceptible to this due to the enhancement
of other lines and information.

**Proposed solution**
Limit the length from the stack trace to 5000 characters that are
provided to AI

## How did you test this change?
1) Found an error with a stack trace more than 5000 characters (`Failed
to execute 'removeChild' on 'Node': The node to be removed is not a
child of this node.`)
2) Ask Harold for help
- [ ] Suggestion is returned
- [ ] Suggestion is still relevant

![Screenshot 2023-07-31 at 10 57 42
AM](https://github.com/highlight/highlight/assets/17744174/c90de356-6dd1-4245-a38d-e0877a1d9625)

## Are there any deployment considerations?
N/A