Moved OpenAI call for session insights to Go backend (#6193)
## Summary

Currently the session insight uses an AWS lambda function to get events
for a session, and then calls the chatgpt API using the nodejs sdk
provided. This is a different approach to the Error Suggestions, which
uses an unofficial Go SDK for OpenAI APIs. This changes moves the
session insight call to Go, so all OpenAI features can be managed in a
single place.

## How did you test this change?

Tried locally and tested on a few sessions

## Are there any deployment considerations?

Depending on how the AI Session Digests are being sent, this could break
the functionality. We should ensure this does not break that
functionality before merging