add lint step to ruby sdk (#6254)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR lints the ruby sdk using
[rubocop](https://github.com/rubocop/rubocop) which is the ubiquitous
ruby linter.

I used `bundle exec rubocop -a` to autocorrect as much as possible. From
there, I used manual correction on everything else (or disabled in
`.rubocop.yml`). For things I wasn't sure on, I added an inline comment
to ignore that specific check.

Since a lot of this PR is formatting,
[remove](https://github.com/highlight/highlight/pull/6254/files?w=1)
whitespace changes when reviewing.

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Build passes

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->