Native SVG components for reflame (#6279)
## Summary

Reflame now has native SVG components! It's API compatible with the
default SVGR configuration that we use here (i.e. components live at the
`ReactComponent` named export), but under the hood it generates a SVG
sprite and lightweight shim components that refer to nodes within the
SVG sprite (more on this approach here:
https://benadam.me/thoughts/react-svg-sprites/) for much better build
performance and runtime performance than SVGR, which renders all SVG
nodes as React nodes.

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?

All the icons in the new preview look identical so far, but let me know
if you notice any irregularities!

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->