avoid writing service.* for frontend logs (#6273)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

We are leaking public graph service name and version into customer's
frontend logs. The reason this is happening is because our OTEL endpoint
serves as the backend for client SDK (i.e. the client SDK calls
pub.highlight.io).

We actually already have similar code that guarded against this for
leaking our public graph's OTEL information:

https://github.com/highlight/highlight/blob/6268-public-graph-service-information-leaking/backend/otel/extract.go#L263-L265

https://github.com/highlight/highlight/blob/6268-public-graph-service-information-leaking/sdk/highlight-go/otel.go#L42-L48

But we didn't update this when we introduced the service notion. This is
also the reason why we were having issues when writing a service record
on ingest (see
[slack](https://highlightcorp.slack.com/archives/C02CJANPHQS/p1691081920973589?thread_ts=1691080057.488759&cid=C02CJANPHQS)).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

* Create a new project locally (it's important for this to be different
from project id = 1)
* Spin up a local react app and use this config:

```js
H.init('<NEW_PROJECT_ID>', {
    backendUrl: 'https://localhost:8082/public',
    tracingOrigins: true,
    networkRecording: {
      enabled: true,
      recordHeadersAndBody: true,
    },
});
```

Observe that we do not have the `service_name` and `service_version`
populated:

![Screenshot_2023-08-08_at_4_12_08_PM](https://github.com/highlight/highlight/assets/58678/103b6136-7e5e-4cc0-97fa-ceff175ffa1c)


## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->