cache GetProject (#6263)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

This PR caches `GetProject` with redis for faster lookups. Since
projects don't typically change, I did set the expiration to one minute.

The motivation here is that I would like restore #6214 but that was
doing many project lookups in a loop. Hence, adding caching will ensure
we don't bottleneck that code path.

Additionally this fixes an issue whereby redis was not cleaning itself
up on test runs (see the first failing build on this PR).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Tests pass

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A