Fix mismatched graphl peer dep version
`subscriptions-transport-ws` has a peer dependency on `graphql` specified as `graphql: ^15.7.2 || ^16.0.0`, which is incompatible with the range `^15.5.0` specified in frontend.

This made the dep tree unresolvable if we adhered to the versions specified in yarn.lock when I tested out the new yarn lockfile sync feature on the highlight repo locally. Bumping to ^15.7.2 resolves the issue.