ensure stacktraces from promises are recorded correctly (#6434)
## Summary

We noticed that we get browser-side errors like
[this](https://app.highlight.io/1/errors/uIzopzx1FAGMCkXv0SxeoGwcK1f0?page=2&query=and%7C%7Cerror_state%2Cis%2COPEN%7C%7Cerror-field_timestamp%2Cbetween_date%2C2023-07-26T18%3A05%3A17.091Z_2023-08-25T18%3A05%3A17.091Z)
recorded without a stacktrace.
The stacktrace is sent from the client as `[]` because we had fallen
back to that value when the
stacktrace parsing library would error. 
That library should only error when a stack trace of `null` or
`undefined` was passed in,
so we should take care of cases where that could be happening.
Otherwise, we should
be explicit about raising the error about a custom-passed stacktrace.

## How did you test this change?


![image](https://github.com/highlight/highlight/assets/1351531/6c5dcd50-4ad9-46e5-bd7a-1fbd8a245506)


## Are there any deployment considerations?

New highlight.run version released.