prefer nearby brokers to avoid cross-az kafka network traffic (#6531)
## Summary

Turns out that we've had the ability to have our consumers prefer the
nearest broker
since we have been on [kafka 2.4
](https://cwiki.apache.org/confluence/display/KAFKA/KIP-392%3A+Allow+consumers+to+fetch+from+closest+replica),
but we have not been using the feature.

This change is based on segment io tickets
https://github.com/segmentio/kafka-go/issues/415 and
https://github.com/segmentio/kafka-go/issues/984 which describe a way to
provide the AZ name to the client
to have the consumer groups prefer the nearest rack (avoiding cross-az
traffic).

## How did you test this change?

Local deploy continues to work.
![Screenshot from 2023-09-05
22-33-31](https://github.com/highlight/highlight/assets/1351531/fee7a086-e635-47a4-ae5b-da900420078b)


Testing the ECS rack preference is challenging locally, but should be
safe to rollback with no data loss (we might just have a small delay in
ingesting data).

## Are there any deployment considerations?

Will be monitoring the release. Only readers should be affected, so
worst case we would have an ingest delay.
Will be following up about billing costs for cross-az traffic.