fix: resolves issue with OTLP endpoint configuration (#6599)
## Summary

Currently the OTLP endpoint can only be configured if it is not on
premise and in a dev or test environment. This is not ideal if the
environment is in docker and designed for production. I see no issue
that the OTLP endpoint can not be changed if provided in the
environment. This change would resolve an issue with #6545 requiring
that it be designated as test or dev to work in helm.

## How did you test this change?

Tested locally in docker and via helm on private EKS

## Are there any deployment considerations?

Consider any potential implications for on premise usage but shouldn't
be impacted since it's utilizing env for any overrides.

---------

Co-authored-by: TheBasedBob <bob@degods.com>