Update GitHub enhancement error logic (#6740)
## Summary
Currently, GitHub stacktraces perform by finding the first 5 files that
do not contain known system level paths. This is hard to manage as
different systems have different paths, and the list is ever growing. If
we are missing a path, it is very likely that the service will be put
into an error state due to too many errors.

Update this logic to:
1. Fetch all files that do not contain known system level paths
2. If a 404 is received from GitHub for a specific file, cache that
result so we do not over fetch on errors
3. Increment an error count only if a file contains no enhancements
4. If this occurs for 5 errors straight within an hour (i.e. different
errors since 404s are cached), then put the service in an error state.

## How did you test this change?
1) Check that errors are still being enhanced
2) Errors with more than 5 enhanceable tracers are all enhanced
3) Files that result in errors are cached and not overfetched from
GitHub


https://www.loom.com/share/6b0dea2a74f947069f0fdf48df3d00c7?sid=268cdb53-8455-4f9b-8479-8ef454b6b47a

## Are there any deployment considerations?
Keep an eye on number of errors and requests in logs

## Does this work require review from our design team?
N/A