ensure billing details are correct to ingest data (#7316)
## Summary

Requires that payment details are correct to allow ingesting data.
Payment detail validity are checked based on the following criteria:
* An uncollectible invoice is present
* An open invoice with failed attempts is present
* The customer has no valid payment methods in stripe

We check these criteria as part of our stripe reporting. If an issue is
detected for the first time,
the customer gets an email and has 7 days to correct the issue. If the
issue is not corrected in that time,
highlight stops ingesting data and shows the billing issue banner and
feed cards.

Closes #7274

## How did you test this change?

https://www.loom.com/share/baa17f8cc8dd484da21fe07840fa17b5

once a payment method is added to the test customer

https://www.loom.com/share/805c9d2c26984ccf92e5d0a12086d57e

![Screenshot from 2023-12-12
11-41-17](https://github.com/highlight/highlight/assets/1351531/e7f2cab4-9beb-409a-8092-ca1bc06840d7)
![Screenshot from 2023-12-12
11-41-21](https://github.com/highlight/highlight/assets/1351531/eea01395-e76d-47c5-a5cb-a3136c93eefd)
![Screenshot from 2023-12-12
11-41-25](https://github.com/highlight/highlight/assets/1351531/7c085bed-3d4a-4331-8d5e-4e4c664ee59b)


## Are there any deployment considerations?

No

## Does this work require review from our design team?

No