Fixed "Suggested Edits" Docs link (#7464)
## Summary

Fix for "Suggest Edits" docs button not linking properly in prod. When
merging the github docs link with the relative docs path, precautions
were taken to prevent double slashes. The links would replace every '//'
with '/'. However, this also applied to the '//' in 'https://', which
broke the link.

## How did you test this change?

Click-testing

## Are there any deployment considerations?

No.

## Does this work require review from our design team?

No.