SessionAlertPage update (#7457)
fixes #7128
@SpennyNDaJets, thought I'd try my hands on this.

## Summary
Update the SessionAlertPage to not update the session form when
alertType changes

## DEMO

https://drive.google.com/file/d/1frNqhyQCxQHo5neDWvLjVS3ALBav9UlW/view?usp=sharing

# Filtering relevant fields
Since we were no longer resetting the form we needed to omit previously
"filled" form data that
is not relevant to the session type being created. I omitted fields
based on the alert configuration.
Alert threshold and threshold window fields are however not omitted. 

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- Visit the new/edit session page and fill in some form like "excluded
environments" or any of the channels to notify
- Change the alert type
- Observe that data in the channels to notify/excluded envs or any other
field that is displayed for the form remains

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
NO

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
NO
<!--
 Request review from julian-highlight / our design team 
-->

---------

Co-authored-by: Jemilu Mohammed <jemilu@crowncitytechnologies.com>