Traces: link to related resources (#7700)
## Summary
Allow users to see the related session, errors, and logs with a trace
from the trace view. These buttons will link the user to the page with
the correct filter, but the user can also view errors and logs in the
tabs below (for now). These tabs are planned to be removed in exchange
for side panel views.

## How did you test this change?
1) View the traces and click on a traces
- [ ] View errors button is disabled if no errors
- [ ] Clicking view errors button links to errors page with correct
filters
- [ ] View session button is disabled if no session
- [ ] Clicking view session button links to session with correct
absolute time
- [ ] View logs is enabled
- [ ] Clicking on view logs button links to logs page with correct
filters


https://www.loom.com/share/217973d784c748a198151ff7d0dbfdfa?sid=50831ec3-3e35-43c1-9b71-c59d87a05a92

## Are there any deployment considerations?
N/A

## Does this work require review from our design team?
N/A