modify password auth to store auth token in cookies to make it accessible across multiple tabs.
I considered just storing the token in localStorage. I was discouraged by the fact that we needed to "manually expire(remove/delete)" the token.

Managing cookies client side without the extra security of limiting it to a particular domain and using "secure" mode is not particularly encouraged but offers a better set of trade-offs against localStorage. I kinda prefer this.