Fix flame graph not appearing on network resources
Updates the behavior of tab content to not be mounted until the tab is shown. This is consistent with the behavior we had in our old `Tabs` component and fixes a bug where we don't calculate the `width` property on the trace flame graph when put behind a tab on the network resource panel.