[SUP-48] support 7 day retention for sessions / errors (#8769)
## Summary
- add `SevenDays` retention option
- change the Stripe logic to not require Stripe prices if an override is
set. Not creating Stripe prices for 7 day retention as this is a custom
agreement.
- load `retention_period` / `errors_retention_period` from the workspace
in the app context - show loading screen until this is loaded
- filter presets to remove any that are larger than the retention
period. For 7 day retention, presets default to 7 days.
- apollo caching improvements to reduce network calls and get rid of
white screen transitions (became a problem after requiring workspace
details to be loaded)
- remove `workspace_for_project` and instead use a nested
`project.workspace`
- create a `Project` fragment to standardize the project fields that are
loaded so new fields don't have to be retrieved. This doesn't affect
performance since all of these fields are already returned from the DB
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- do not have to make any changes in Stripe
- will run `update workspaces set retention_period = 'SevenDays' where
id = x` after this is deployed
- have already added S3 lifecycle rules for the `SevenDays` option
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?

<!--
 Request review from julian-highlight / our design team 
-->