Ensure session feed query pulls in the latest sessions (#8821)
## Summary

We are seeing an issue where new sessions aren't appearing in the feed
because of some logic we added to help with showing cached results
immediately on page load. This fixes the issue by extending the end date
time into the future before subtracting from it for rounding. We only do
this when the user is trying to show real time results.

## How did you test this change?

Click tested and checked the values using a debugger when using a fixed
time and a relative time.

## Are there any deployment considerations?

N/A

## Does this work require review from our design team?

N/A