Version Packages (#9167)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## highlight.run@9.3.0

### Minor Changes

-   0a8a9ffdc: add cookie session persistence

### Patch Changes

- d2e00028a: correctly set x-highlight-request on outgoing fetch/xhr
requests with duplicate tab recording.
the sessionID in the x-highlight-request would not be set correctly
after recent changes
corrected the multi-tab behavior to clear the local storage sessionID
value to ensure
    new tabs started unique sessions. corrects bug affecting >=9.1.5
-   0a8a9ffdc: update otel webjs network span naming

## @highlight-run/next@7.5.21

### Patch Changes

-   Updated dependencies [d2e00028a]
-   Updated dependencies [0a8a9ffdc]
-   Updated dependencies [0a8a9ffdc]
    -   highlight.run@9.3.0
    -   @highlight-run/node@3.9.0
    -   @highlight-run/react@5.0.0

## @highlight-run/react@5.0.0

### Patch Changes

-   Updated dependencies [d2e00028a]
-   Updated dependencies [0a8a9ffdc]
-   Updated dependencies [0a8a9ffdc]
    -   highlight.run@9.3.0

## @highlight-run/remix@2.0.52

### Patch Changes

-   Updated dependencies [d2e00028a]
-   Updated dependencies [0a8a9ffdc]
-   Updated dependencies [0a8a9ffdc]
    -   highlight.run@9.3.0
    -   @highlight-run/node@3.9.0
    -   @highlight-run/react@5.0.0

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>