Address some Next.js build warnings
Addresses the following warnings we were seeing in our build logs:

```sh
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.
```

```sh
59:6  Warning: Using `<img>` could result in slower LCP and higher bandwidth. Consider using `<Image />` from `next/image` to automatically optimize images. This may incur additional usage or cost from your provider. See: https://nextjs.org/docs/messages/no-img-element  @next/next/no-img-element
```