Add a python logging OTEL quickstart guide (#9735)
## Summary

Adds a simple python logging OTEL quickstart guide.

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?

n/a

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

n/a

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?

n/a
<!--
 Request review from julian-highlight / our design team 
-->