move polling logic up a level, add useGraphTime helper (#9729)
## Summary
- move polling logic outside of `Graph` and make `Graph` update on
start/end date changes - this helps by polling for all graphs at the
same time to eliminate some of the jankiness
- add `useGraphTime` helper that can be used across different pages that
show graphs
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally on dashboard, graphing editor, and alert pages
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- worked off the `zane/ui_memo` branch as that had some structural
changes and I didn't want to fix conflicts. will merge that branch in
before this one
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->