add missing fields to mvs (#9923)
## Summary
- there were no MVs to write these attributes to the `trace_key_values`
table, so they're missing from key and value suggestions
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- checked a couple fields (e.g. `telemetry.distro.name`) after running
migration script locally
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->