create dashboards/alerts from logs/traces page (#9930)
## Summary
- support creating dashboards/alerts from logs/traces page
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- @julian-highlight - I thought it looked a little janky to hide the
button by default and only show it when a query has been entered, so I'm
showing it regardless. Lmk what you think!
<!--
 Request review from julian-highlight / our design team 
-->