Zane/main queue tuning (#10046)
## Summary
- consumer for the main topic seems to be saturating at 1000 and could
be a bottleneck (low bandwidth reading from kafka)
![Screenshot 2025-03-04 at 1 43
05 PM](https://github.com/user-attachments/assets/47e7c817-cf48-4234-acc6-0eda269daa4a)

- add a `MainQueueSize` parameter to `system_configurations` to make
this tunable
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- built locally, will observe performance in prod
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->