Replayer performance optimizations (#10163)
## Summary

A few optimizations to the session replayer:

* Changes how rrweb replaces actions by completely replacing the array
rather than using `splice` to reset it. This seems to fix an infinite
loop we were seeing, though probably has some tradeoffs with creating a
new array rather than reusing the existing.
* Filters out custom events before passing events to rrweb since those
are only used on the timeline, not for replay.
* Adds some mechanisms for creating many track events at once and
viewing very large DOM trees to test those scenarios in
recording/replay.
* Prevents Stripe from loading when any page is rendered in our app and
lazy loads it only when needed on the update plan page.

## How did you test this change?

Click tested locally and in the Reflame preview, which included the
rrweb updates. Compared prod session viewing to the app preview for some
sessions that had obvious playback issues.

## Are there any deployment considerations?

* Need to release https://github.com/highlight/rrweb/pull/122 and update
rrweb in this PR before merging.
* Will need to follow up with customers to let them know replay should
be better (see Jira ticket for details on who to contact).

## Does this work require review from our design team?

N/A