use correct retention date for errors (#10221)
## Summary
- we're using the incorrect retention date when querying for errors,
noticeable for a customer with sessions retention = 7 days and errors
retention = 6 months
- patched from LD repo
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- tested in other branch, will validate customer issue after deploying 
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->

Signed-off-by: Zane Mayberry <zmayberry@launchdarkly.com>